fix: Enhance AI assistant and SEO metadata
Improve the AI assistant's persona and capabilities by expanding its system instructions. Update the website's meta description and JSON metadata to accurately reflect the company's offerings and location in Nairobi. Refine product image loading for better fallback handling. Add new view states for privacy and terms pages to the types definition.

+++ server/src/entities/Expense.ts
@@ -1,4 +1,3 @@
-
 import { Entity, PrimaryGeneratedColumn, Column, CreateDateColumn, ManyToOne } from 'typeorm';
 import { User } from './User';
 import { Branch } from './Branch';
@@ -22,7 +21,7 @@ export class Expense {
   @Column()
   title!: string;
 
-  @Column('decimal', { precision: 10, scale: 2 })
+  @Column('decimal', { precision: 16, scale: 2 })
   amount!: number;
 
   @Column({
